ci(js): fix npm 12 publish flags; dart: fix analyze warnings, bump version for pub.dev release
npm 12 throws on the unrecognized --no-git-checks flag and now requires
--tag for prerelease publishes; drop the former and add the latter to
the dry-publish step.

pub.dev's automated publishing requires pubspec.yaml's committed version
to match the pushed tag (no CI-side version injection path exists,
unlike the other language publishers), so bump it and the changelog
here as the one documented exception. Also fixes the two dart analyze
warnings that were failing pub.dev's non-interactive publish dry-run.

+++ dart/ipaddress/lib/IPAddress.dart
@@ -1278,7 +1278,7 @@ class IPAddress {
     for (var i = num; i < this.ip_bits.bits; i++) {
       var a = log(i) ~/ log(2);
       if (a == log(i) / log(2)) {
-        return this.prefix.add(a as int);
+        return this.prefix.add(a);
       }
     }
     return Result.error("newprefix not found ${num},${this.ip_bits.bits}");
